returnValueHasChanged() method for method-related events
