+++ src/Event/ProxyableMethodCalledEventInterface.php
@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 namespace Dealroadshow\K8S\Framework\Event;
 
 interface ProxyableMethodCalledEventInterface extends ProxyableMethodEventInterface
